Compiler architecture update Design Flow Manager update  Design flow manager has been updated to avoid use of boost::filtered_graph.  Design flow step signature is now an integer instead of a string to improve performance.  Design flow manager profiling has been improved; CSV file generation has been added. Tree manager update  Tree manager is now handling tree_node instances by default instead of tree_reindex to avoid multiple indirection.  Tree reindex remove has been implemented to remove all tree_reindex instances after Bambu IR parsing is completed.  GET_NODE, GET_INDEX_NODE have been removed. Abseil update to version 20240116.2 included as submodule Parameter class template refactoring to improve string to type casting performance General code refactoring to reduce the usage of dynamic_cast Removed unused tree node memory_tag Pragma analysis step refactoring Bit value opt step refactoring Other minor fixes
Minor refactoring

Hide DesignFlowStepPrioritySet implementation

+++ src/SpiderParameter.cpp
@@ -381,22 +381,22 @@ void SpiderParameter::CheckParameters()
                break;
          }
       }
-      setOption(OPT_input_format, static_cast<int>(input_format));
+      setOption(OPT_input_format, input_format);
    }
    else
    {
       Parameters_FileFormat input_format = GetFileFormat(getOption<std::string>(OPT_input_format), true);
-      setOption(OPT_input_format, static_cast<int>(input_format));
+      setOption(OPT_input_format, input_format);
    }
    if(!isOption(OPT_output_format))
    {
       Parameters_FileFormat output_format = GetFileFormat(getOption<std::string>(OPT_output_file), false);
-      setOption(OPT_output_format, static_cast<int>(output_format));
+      setOption(OPT_output_format, output_format);
    }
    else
    {
       Parameters_FileFormat output_format = GetFileFormat(getOption<std::string>(OPT_output_format), false);
-      setOption(OPT_output_format, static_cast<int>(output_format));
+      setOption(OPT_output_format, output_format);
    }
 #if HAVE_TECHNOLOGY_BUILT
    for(const auto& input_file : getOption<CustomSet<std::string>>(OPT_input_file))
